SI-6654 & SI-7005 - Map#filterKeys and Map#mapValues not serializable
Fixes SI-6654
Fixes SI-7005

Review by @axel22
Review by @rkuhn
