+++ src/library/scala/collection/MapLike.scala
@@ -227,7 +227,7 @@ self =>
   def default(key: A): B =
     throw new NoSuchElementException("key not found: " + key)
 
-  protected class FilteredKeys(p: A => Boolean) extends AbstractMap[A, B] with DefaultMap[A, B] {
+  protected class FilteredKeys(p: A => Boolean) extends AbstractMap[A, B] with DefaultMap[A, B] with Serializable {
     override def foreach[C](f: ((A, B)) => C): Unit = for (kv <- self) if (p(kv._1)) f(kv)
     def iterator = self.iterator.filter(kv => p(kv._1))
     override def contains(key: A) = self.contains(key) && p(key)
@@ -241,7 +241,7 @@ self =>
    */
   def filterKeys(p: A => Boolean): Map[A, B] = new FilteredKeys(p)
 
-  protected class MappedValues[C](f: B => C) extends AbstractMap[A, C] with DefaultMap[A, C] {
+  protected class MappedValues[C](f: B => C) extends AbstractMap[A, C] with DefaultMap[A, C] with Serializable {
     override def foreach[D](g: ((A, C)) => D): Unit = for ((k, v) <- self) g((k, f(v)))
     def iterator = for ((k, v) <- self.iterator) yield (k, f(v))
     override def size = self.size
